+++ Examples/minimum_user_for_s2e/src/src_user/IfWrapper/Sils/uart_sils.cpp
@@ -3,7 +3,7 @@
 #include "../../Settings/port_config.h"
 
 #ifdef USE_SCI_COM_UART
-#include "uart_sils_sci_if.h"
+#include "uart_sils_sci_if.hpp"
 #endif
 
 int OBC_C2A_SendFromObc(int port_id, unsigned char* buffer, int offset, int count);
