+++ Examples/minimum_user_for_s2e/src/src_user/IfWrapper/Sils/uart_sils_sci_if.cpp
@@ -6,7 +6,7 @@
             ccsds_sils_sci_if.c/hのほぼコピー
  */
 
-#include "uart_sils_sci_if.h"
+#include "uart_sils_sci_if.hpp"
 
 
 // 最初だけ初期化して、プログラム終了時にポートを閉じるようにしたい
